+++ pg_documentdb/src/aggregation/bson_aggregation_nested_pipeline.c
@@ -574,6 +574,8 @@ HandleDocumentsStage(const bson_value_t *existingValue, Query *query,
 	PgbsonWriterInit(&writer);
 
 	ParseAggregationExpressionContext parseContext = { 0 };
+	parseContext.collationString = context->collationString;
+
 	AggregationExpressionData expressionData;
 	memset(&expressionData, 0, sizeof(AggregationExpressionData));