+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Support `$convert` on `binData` to `binData`, `string` to `binData` and `binData` to `string` (except with `format: auto`) *[Feature]*
 * Fix list_databases for databases with size > 2 GB *[Bugfix]* (#119)
 * Support ARM64 architecture when building docker container *[Preview]*
+* Support collation with `$documents` and `$replceWith` stage of the aggregation pipeline *[Feature]*
 
 ### documentdb v0.102-0 (March 26, 2025) ###
 * Support index pushdown for vector search queries *[Bugfix]*